+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 * This project follows the guidelines outlined on [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+- Pop serial buffer in single call to prevent race conditions
+- Instead of multiprocessing.Queue use the more generic queue.Queue
+- Catch Marlin Resend messages
 
 ## [0.22.3] - 2016-07-17
 ## Fixed